SelfRefVecItem trait is defined
* swap and truncate methods are added to the trait.
* SelfRefVecItem trait is defined for prev-next type of relationships, to be extended later for different kinds of relations.
